Permission resource create and list done

+++ database/migrations/2022_11_04_160901_create_permissions_table.php
@@ -15,7 +15,7 @@ class CreatePermissionsTable extends Migration
     {
         Schema::create('permissions', function (Blueprint $table) {
             $table->id('permission_id');
-            $table->unsignedBigInteger('role_id');
+            $table->integer('role_id');
             $table->text('permission');
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
             $table->timestamps();
